refactor  /utilities/auth.mjs to /auth/index.mjs
